Remove directory-schema
